Avoid ClassNotFoundException when using Community Edition
